Refactor sensors (state/config) and rule condition handling
Sensors are now derived from abstract 'Resource' class which holds
arbitrary resource items. These items can be queried, are type checked,
range checked (validation) and provide timestamps for 'set' and 'change'.

This should be easier to maintain. Rule processing is optimized for speed
and is coupled to the Resource system.