IKEA motion sensor config.durarion and state.dark
Fix `config.duration` and implement `state.dark` for the IKEA motion
sensor.
- `config.duration` is updated (only) when presence is detected from
the dial setting on the sensor.  Only when the value has changed, a web
socket notification for config is sent;
- `config.duration` is read-only (cannot be updated through the REST
API).
- `state.dark` is updated when presence is detected.  When the dial is
set to Day, it’s value is true; when the dial is set to Night, it’s
value reflects the motion sensor’s measured light level.
- The `state.any_on` of the linked group and `state.on` of the lights
in that group are longer set to `true` when `state.dark` is false,
while detecting presence.