Support Xiaomi vibration sensor sensitivity read/write
    "config": {
        "battery": 98,
        "on": true,
        "pending": [],
        "reachable": true,
        "sensitivity": 1,
        "sensitivitymax": 21,
        "temperature": 3100
    }

Sensitivity is from 1..21 where 1 is the most sensitive.

When a PUT on config/sensitivity is done it will be written as soon as either
the sensor button is pressed or the hourly report is recived. Further the
configured value will be verified and autocorrected each hour.

Issue https://github.com/dresden-elektronik/deconz-rest-plugin/issues/1025