Add support for ZigBee Green Power (ZGP) mesh proxy handling (not working, wip)
ZGP Proxies such as hue lights can forward ZGP messages from Friends of Hue
and Hue Tap switches through the network. When joining such a device the
proxies need to be commissioned.

This commit adds the basics of the procedure but isn't working yet since the
GPD key needs to be decrypted.