Add CLIPDaylightOffset sensor (wip)
The sensor is ment to represent arbitrarly points relative to a daylight
specific point in time, which can be used in rules.

config.mode: The reference point, like sunrise or sunset.
config.offset: The offset to the reference in seconds.

state.localtime: The daily calculated localtime of the offset. Rules can
                 reference it in conditions, e.g. a rule 'in' condition
                 to specify activity between midnight and 3 hours before
                 sunrise.

```json
{
    "name": "Sun Offset Ref P1",
    "type": "CLIPDaylightOffset",
    "manufacturername": "Phoscon",
    "modelid": "DLOFFSET01",
    "swversion": "1.0",
    "uniqueid": "00:21:2e:ff:ff:03:d4:9a-03-4e32",
    "config": {
        "on": true,
        "mode": "sunset",
        "offset": 121
    },
    "state": {
        "localtime": "2019-08-28T04:03:04",
        "lastupdated": "2019-08-28T00:05:53"
    }
}
```