Patch invalid node descriptors
This is intended to cover rare cases where deconz core determines an invalid node descriptor due to mac capabilities being 0x00 via .isNull()-Function. This happens, if the allocate short address bit is not set, although it complies with the standard (see 07-5123-06-zigbee-cluster-library-specification.pdf, 2.3.2.3.6 MAC Capability Flags Field).
After successful pairing and light/sensor creation, the ND in node cache is 0x00000000000000000000000000 and therefore, a corresponding entry for ZDO endpoint is not written to the database. In those cases, after a deconz restart, the ND is updated wrongly, as it always assumes the ND from the previous node (from a node cache perspective: device internal ID i=x, ND assigned i=x-1). It is unknown if that's a glitch or on purpose. This is especially inconvenient if the device uses manufacturer specific attributes, clusters or commands.
The proposed code validates a ND received from a node descriptor response and patches it with the allocate short address bit, therefore becoming valid again. It then updates the node cache and forces a DB update. It also patches the manufacturer code for Develco devices running older firmwares.
This is primarily applicable for certain Develco devices, but may also apply Bosh, Ubisys and BJ devices.