Fix network permit join set after deCONZ start
This could happen when permit join value > 0 was stored in the database and
restored after application startup. In my setup this caused invalid light nodes created from devices which where actually sensors and switches.

This PR removes database store/restore of the permit join value which must be set
to 0 on each startup and can only be set by REST-API.