Add StateChange helper class for fault tolerance
This class carries out commands and verifies the result it can be used for routers and sleeping devices. It is fault tolerant by verifying the result.

If the new state wasn't reached the command will be send again. This repeats until a timeout is hit, which can be arbitrarily long.