Relax IKEA lights reporting configuration
The initial approach to use the settings from the IKEA gateway results in a huge amount of rapid reports during state transitions. The lights are sending reports for every tiny change especially for Level and Color  Control clusters.
This commit ensures that there are at least 5 seconds between two reports for an attribute. For color temperature lights reporting for ColorXY attributes is disabled (these need more logic to be handled properly).