Improve setup of Aqara Opple switches
This PR ensures the magic attribute to enable button events is written in the hourly report, and at least once per deCONZ start.
After startup writing the attribute can be enforced by pressing the 'C' button on the back of the switch.
There was also another issue when the switch is paired a simple descriptor request loop could happen, this is fixed in deCONZ core now.

 Issue https://github.com/dresden-elektronik/deconz-rest-plugin/issues/3670