Align VOC concentration on Sensirion levels for air quality
Previously, the scaling has been aligned with Develco documentation for Develco AQSZB-110 and Bosch Air sensors. As this has apparently been revised and also taking the referenced documentation into account, the levels are now aligned on German Federal Environmental Agency.

Based on that, the sensor measurements now appear to be more reasonable.