Fixed issue with the Eria Adurosmart Wireless Dimming Switch sending an event for Button 1 for any button pressed.
This switch has a situation where it sends the same endpoint/cluster/command ID for every button pressed and only makes the distinction between buttons via payload value.

The current logic in `DeRestPluginPrivate::checkSensorButtonEvent()` currently does not support this type of distinction, and the button map (field #5) only seems to support the first digit of the payload value and not the full payload value.

Therefore all of the information that distinguishes between the buttons ends up being ignored and the button events are being sent as the first mapped entry (button 1).

This commit changes the logic within `DeRestPluginPrivate::checkSensorButtonEvent()` to allow the value of the button pressed to be modified once a valid mapped entry has been found, and adds specific logic for the Eria Adurosmart Wireless Dimming Switch to change the button value based off of the payload value.