Change Lidl/Silvercrest Smart Wireless Door Bell from ZHASwitch to ZHAAlarm resource
This device is technically an alarm sensor and should be treated as such. By doing that, all assiciated IAS attributes will be exposed as supported by the device. Any triggering in automation systems must then be done on the `alert` resource item.

For this take to take effect in environments where the device is already in use, it must be reset and the corresponding ZHASwitch resource under `/sensors` should be manually deleted via API. prior to re-pairing.