Fix IKEA lights not turn off in scenes and IKEA unicast off commands
IKEA firmware has a bug that it ignores the off setting in a scene. As workaround unicast Off With Effect commands are send to IKEA lights which are off in a scene.

In same fashion plain off commands to IKEA lights are changed to use Off With Effect command.

The group recall scene code was refactored to be more dense and properly handle color loop.