Raise short database save delay from 5 seconds to 1 minute
Writing to the SD card every 5 seconds due lights and sensor updates blocks the whole application way too often, sometimes for seconds.

Since the state is changed on clean deCONZ shutdown anyway the only drawback of the increased delay is the loss of 1 minute data in case of power failure.