Use QLatin1String for all HTTP header queries
This is preparation for a refactored QHttpRequestHeader class which uses static strings where ever possible.

A call like `hasKey("If-None-Match")` implicitly creates a QString since the method expects a `const QString&`.  This leads to a operator new call in QString. Since we mostly just have static strings this can be avoided alltogether.