Update QList iterators and prevent copy of Simple Descriptor
In v2.12.0 various uses of QList in the deCONZ library will be replaced with std::vector. QList is way slower than std::vector and more badly has the habbit of detaching an allocating extra memory when iterated over a non const container.

The removed copying of the Simple Descriptor has an even bigger impact since this was done very frequently and caused needless copies of all clusters and attributes going into the millions after just 10 minutes. With this PR only a const reference into the object in core is used.