Prevent continuous write attempts on Hue specific attributes
With the previous code changes, setting the already set values once again for ledindication, suertest, sensitivity and devicemode results in repeated write attempts for appr. 7200 secs, until the next attribute read. The expected attribute report doesn't occur so that the write is considered unsuccessful as the corresponding `RConfigPending` flag is not cleared.

Now, the next write attempt is conducted upon the next attribute report, but no longer than 7200 secs.