Add EventEmitter class to decouple enqueueEvent() from plugin
For now with a global `enqueueEvent()` free function, to not change hundrets lines of code. In future we can make this cleaner and get rid of the globals.

One timer call handles multiple events, if the total processing time is below 10 ms. This is done because one timer activation might take +-10 ms until called by Qt (borrowed from Device code).