Fix: Make Xiaomi WRS-R02 switch configuration more robust
As there's always one more place to whitelist, the device now adds `RConfigPending` after restart to allow switch configuration.

Additionally, to make the configuration procedure more robust, `RConfigPending` is now set to require configuration upon device announcement (currently applicable for Opples and WRS-R02). This ensures the Xiaomi device operation mode is marked for writing when the device is reset AND a ZHASwitch resource already exists which is not marked as deleted. Without it, the device does not get configured if reset more than once while deconz is running.