Enable APS ACKs for ZCL Default Response
For end devices it can happen that the APS request is enqueued in the firmware but never sent, in which case the APS confirm ends with a timout status code (0xF0).

It looks like a timing issue, where the plugin doesn't push the APS request fast enough, and the end device MAC data polls too fast for a response. Or when the end device initial request required an APS ACK from the coordinator.

This can be seen for Samjin Motion sensor, where not a single ZCL Default Response made it to the device. I suspect a few other devices could be affected as well.

The fix is simple, with APS ACK enabled the firmware will just retry to send the response if it doesn't work on the first attempt.

This is more heavy, but given that the ZCL Default Response is kind of important in order to keep devices alive, worth it.