Only queue IAS presence restoration if needed
Add support for checking whether a sensor supports sending reports
on alarm restoration by checking the `Restore report` bit of the IAS
ZoneStatus and against a list of sensors known to send restore reports.
Update IAS Zone presence handling to only queue a reset for sensors that
do not send reports on restore.