Fix Ikea switches for new and old firmware
The new Ikea firmware broke group cast support. This PR sends an Identify Query Response during joining to the switch.

With that the switch automatically sends commands as unicast to the coordinator, which it didn't before.

No extra unicast bindings are created, and group bindings are still supported for older firmware.

To prevent duplicated button events, the ZCL sequence number check is enabled for Ikea switches.

Testet with:
- On/off switch
- 5 button remote control
- Styrbar remote control
- Shortcut button