DDF support allocating device groups
For each "auto" entry in the comma seperated group list a new Group is allocated and assigned to the device via device membership and  group.uniqueid.

The group.uniqueid is generated based on the position in the config.group list:

     "auto":            00:11:22:33:44:55:66:77

     "auto,auto":       00:11:22:33:44:55:66:77
                        00:11:22:33:44:55:66:77-1

     "auto,auto,auto":  00:11:22:33:44:55:66:77
                        00:11:22:33:44:55:66:77-1
                        00:11:22:33:44:55:66:77-2

Where 00:11:22:33:44:55:66:77 is the MAC address of the device.
Note: The first group has no "-0" suffix, to be backward compatible with old code.