Enable Gold status DDF files by default
The "Disabled" DDf mode is replaced by "Basic" mode in which only DDF files with enabled status are loaded.

The default is Gold, but Bronze and Silver can also be enabled. This setting is persistent. Strict and Hybrid (formerly Normal) modes are not persistent and revert to Basic mode after each deCONZ start to prevent user errors.

When the modes are changed in Control panel, all Devices with DDF will be reloaded, one per second, to reflect the changes without restarting deCONZ.