Prevent DDF devices to poll too fast in idle mode
This is only a rough workaround which will be replaced by a more precise method later. The problem is in a network with lots of DDF devices the APS queue can be stressed too fast, since the DeviceTick triggers polling without waiting if the previous polling device is already finished.