DDF for OWON THS317-ET Multisensor (temp probe support)
This is a DDF I created for this sensor with the help of @Smanar and @Mimiix. It actually works pretty well for reading the temperature cluster, the battery and the sw version, but this sensor is supposed to report humidity and is also a PIR sensor.
As you can see on the screenshots provided in issue #5738 it probably lack some cluster for these other features to be supported by deCONZ so if someone have a sniffer and this device, feel free to contribute to make other commands usable.
As requested the DDF has been set up to Bronze status.