DDF initial support for top level "matchexpr" property
This allows to select a DDF not only by modelid and manufacturer name
but optionally also other aspects as endpoints of the device.

The initial version supports R.endpoints which returnes an array of active endpoints of the node
e.g. [1, 2]. So it can be queried for specific endpoints or simply the number as:

{ "matchexpr": "R.endpoints.length === 2" }

Note this should only be used in very specific cases!