DDF support sub device "meta" objects
In "meta" JSON objects can be provided which may describe static aspects of a device
not fitting in general "items" or other keywords.

One example is `group.endpoints` of an upcoming PR to describe which endpoints
reflect the index of `config.group`.

{ "meta": {"group.endpoints": [1, 2] } }

The C++ code can query the meta object for a sub device as QVariantMap.
Warning: this is absolutely not meant to add endless things to the meta object, most often it's better to discuss and find a general approach!