Button handler skip some modelid checks; improve sensor.mode
If a device only contains one sensor like ZHASwitch, skip modelid and manufacturer
name checks to find a better sensor.

The button_maps.json contain a mode field which may or may not be set properly in
a sensor. The default was ModeTwoGroups and now is ModeScenes due another PR.

During initial button map fetch check if the complete button map only uses one mode
and correct it in the sensor.mode if needed.