Fix processing rules referencing DDF devices
Since DDF devices and their light and sensor resources are created async, the rule triggers can't be indexed in the plugin constructor.

The PR delays/debounces the indexing so that all devices are handled properly. This is also a bit faster since prior for each device useless container calls where made during startup.