Followup LIDL/Silvercrest door bell (HG06668)
The DDF didn't load due 0-100% "invalid utf8" in config/battery. It's actually valid but needs another fix...

Main problem is that we have different modelid/manufacturernames in the databases out there.

The variants are added so the DDF is picked up for each of these. Note for later we need to have something like:

"apimodelid"
"apimanufacturername"

To express what should actually be exposed and be backward compatible, since the legacy code sometimes replaces the original Zigbee values (Lidl/Tuya).