Update ikea-ota-download.py to work with pyservice
A colleague was trying to update Home Assistant OTA files with this script. We tried using pyscript, but I needed to modify urlopen and urlretrieve to be asynchronous functions for it to work. Similarly I needed to replace print statements with logging in order to get feedback from the script into Home Assistant's general logs. I am not sure if this is the intended use of the script, but this was at least a solution to my colleague's problem that may benefit others.