Skip legacy code when pairing via DDF
While Permit Join is enabled, always wait until manufacturer name and modelid are queried by the Device code. This happens also for non DDF devices.

For lights and sensors which have a DDF, skip legacy code for creating resources, when the DDF status is enabled. This fixes creation of invalid lights.

Note: This change is a bit bold and may show a few missing parts in DDFs but, beside the mentioned fix also speeds up DDF pairing.