Add initial DDF for Tuya smart knob (_TZ3000_4fjiwweb)
It's a weird device having two modes, 'scenes' and 'remote', which can
be toggled by pressing the knob 3 times very quickly (less than 0.5 seconds).

Depending on the mode following button events are emitted:

Scenes:
- 1002 single press
- 1030 rotate clockwise
- 1031 rotate counter-clockwise

The `state.angle` gives the rotation angle 0-360 degrees.

Remote:
- 1002 single press
- 1004 double press
- 1030 rotate clockwise
- 1031 rotate counter-clockwise

In this mode `state.angle` isn't updated.

TODO there are further functions available when rotating the knob while the
button is pressed, these are not implemented yet.

This PR also add support to parse any ZCL command in DDFs and exposed the `ClusterId` to the Javascript engine.

The database is cleaned up from older invalid light resources for this device.