Fix Device code async state enter
This was a tricky one. A state handler function must not process any events before the `state.enter` event arrives. Since this event is delivered async, to not blow up the stack if another state is entered from this event, we could get into troubles and invalid state if other events arrive before the enter event was handled.

- The PR processes all state.enter events (marked as urgent) before any other events.
- Removed the dead code for start/stop timer events.
- The event queue drops duplicated events now, so we don't have multiple awake events and similar floating around.
- Speedup of processing events directly after an APS request is processed. No need to wait for the idle timer.