Fix Tuya time sync in early pairing stages
The time sync could fail if no sensor or light resources where created at the time of the request. This can happen for DDF devices, I've seen it with Nous E6 sensor.

The PR puts the time sync code on top as it is independed of any sensor and light resources.

Related: https://github.com/dresden-elektronik/deconz-rest-plugin/pull/6064