Remove rescaling of consumption in Develco/friend ZHEMI101 DDF
This commit removes the division by 1000 when extracting the consumption
value for the Develco/friend external meter interface ZHEMI101.
This sensor comes in different flavors that is compatible with different
standards. For example the EU version is compatible with the EU standard
of 1000 pulses per kWh, so the value is already correct from the sensor.