Initial Duktape JS engine implementation
This is only a test branch, all API extensions from Qt JS engine are supported.

Note: Some DDFs needed to be adjusted to use `var` instead of `let`.
The engine requires ca. 200KB of RAM to run DDF scripts.

Execution time in release build is 2-3 ms on desktop but can be optimized further.