Fix too frequent database writes (#6487)
When a resource item is changed by DDF code is written to the database quickly. The PR throttles this so that `state/*` items are only written if they don't exist or the version in the database is older than 10 minutes.

In future we need to denote items via JSON to have a storage interval.
There are many items which aren't that important and are ok to
be stored in larger intervals or only when deCONZ is closed.

The PR also fixes that static modelid and manufacturername items are
respected in the DDF init code.