Refactor gateway password code
The main goal was to dynamically load libcrypt with `dlopen`.
The release build on Debian buster doesn't work Ubuntu Jammy, because it has a different version of the library.

Further the code was separated from `DeRestPluginPrivate` into the `crypto/password.cpp` module not using Qt and be testable.

TODO: Move away from libcrypt and use the scrypt functionality which we
already have on each platform.