DDF for Elko Smart ZB Thermostat 16 A (#7225)
There is some general information here:
https://proff.elko.no/elko-smart/ELKO%20-%20Smart%20ZB%20Thermostat%2016%20A_Device%20User%20Guide/English/ELKO%20-%20Smart%20ZB%20Thermostat%2016%20A_Device%20User%20Guide_0000930613.html

Thermostat with support for two temperature sensors and some metering.
There is also the option of setting a schedule, which is ignored for now.
With this DDF, temperature(s) are read, setting temperature and reading
heating state works.

Done-with: Smanar <aurelien.pra@laposte.net>