Create lumi.switch.b1laus01.json
Create a DDF for `lumi.switch.b1laus01`, which is the US no-neutral Aqara switch. It's currently missing a couple config options, namely:

- Flip LED status indicator
- Switch "mode" between "quick" and "anti-flicker"

Both are exposed in Z2M, though I'm not familiar enough with the DDF structure to implement those endpoints. This does make the switch work, however, so that's nice.