Remove legacy FLS-PP (version 1)  color XY quirk
The device is the old black one, not the white FLS-PP lp.
Since the firmware didn't support XY color mode there was a loop in the idleTimer handler iterating over all lights to find out if this device is present. Since the device is likely not in use anymore we can get rid of this CPU cycle eater. If needed introduce a DDF later on.