Optimize updatedLightNodeEndpoint()
This functions loops over all lights on each NodeEvent, which is wasteful on the CPU since only in case of a simple descriptor response the body is executed. Therefore only do this if the NodeEvent actually matches and bail out early otherwise.

The function can likely be removed all together, but needs some testing due the .rx() call.
The function