Relax polling of DDF devices without refresh.interval
For items which don't specify refresh.interval in DDF or at least the generic item, relax to poll every 30 seconds instead of "always and as fast as possible".
This is not ideal and we should have a closer look at the items, also if they use reasonable poll interval in combination with ZCL attribute reporting.