Fix OSRAM Classic A60 RGBW crash
The light does crash when ZCL attribute reporting is configured and the binding table is queried. Likely resulting in a watchdog reboot after which the lights turns on in full brightness.

The fix here is to use polling instead of attribute reporting as well as to disable query of the binding table.

Related https://github.com/dresden-elektronik/deconz-rest-plugin/issues/7546