Philips Hue ZLL/ZB3 match DDF only for newer firmware
The existing DDFs require that the lights have a recent firmware which provides the 0xFC03 cluster.

With this PR the DDFs aren't matched any longer for older firmware lights and fall back to the legacy code.

Later on we can introduce DDFs which also handle the old firmware with standard clusters instead of 0xFC03 cluster.