Add CJ JSON parser (#7722)
CJ is a very fast low level zero allocation JSON parser.
Used instead of Qt and ArduinoJson parsers for new code to parse DDFs and bundles.

Currently there is no C++ interface, the C API is a bit clunky but that way it is as fast as it can get with non SIMD parsers. A later C++ wrapper will be made compatible with the Qt parser since that is used mostly in REST-API code.