Additional special effects for Hue lights (#7922)
With the 1.116.3 / 67.116.3 firmware upgrade from May 2024, most Hue lights get additional special effects (see https://www.philips-hue.com/en-us/support/release-notes/lamps). There's also a new (undocumented) "sunset" effect, which basically is the reverse of the "sunrise" effect.

Most of these changes should be reflected by the API automatically when, the attributes from the Hue Effects cluster are re-read after the firmware update.

This PR adds support for the new "sunset" effect. It also brings the GUI up to date with the API ("opal" and "glisten" were missing and "prism" was still called "loop"). It enables the Hue-specific Startup Current X and Startup Current Y attributes, now that the GUI bug handling attributes with the same ID but different manufacturer codes has been fixed.
Finally, it removes the values for state/effect from the DDFs for the Hue lights, as these are exposed through capabilities/color/effects, and as they depend on the firmware revision.

While changing the C++ code, I also fixed the handling of special effects for (white and/or white ambiance) lights that don't support colorloop, see #7451.