Fix ResourceItems no longer handled when additional DDFs are loaded
Problem in the issue was reproduced by joining a device without DDF, existing devices stopped processing some of the resource items. The main issue was that when loading new DDFs during a session the load counter got updated for all items in all DDFs, but the device items in memory still pointed to former load counter.

TLTR the load counter should only be incremented for the actually new DDF which is loaded into memory.

https://github.com/dresden-elektronik/deconz-rest-plugin/issues/7935