DDF for Third Reality contact sensor (3RDS17BZ) (#8039)
Add a DDF for the Third Reality contact sensor. The DDF is based on the
Ikea Parasoll DDF, as this device also uses the IAS (alarm system) cluster
to report opening/closing events.

The device doesn't seem to support reporting for the Battery Percentage
Remaining, giving an Unreportable Attribute error when trying to configure
it, so this DDF does not configure reporting.

There is an attribute for `config/delay`, which uses a
manufacturer-specific cluster to set the delay before an opening event is
reported.