Temporary solution for _TZ3000_qeuvnohg
This PR aims to override the PR at https://github.com/dresden-elektronik/deconz-rest-plugin/pull/7012, to make these MAINS device(s) available, but also to mitigate the danger inherent to the MAINS type in that PR.

This is a temporary solution until Mains switches are handled better in Phoscon GUI.

Note that in above discussion it was argued, that security is the job for the API client. Since Phoscon GUI does not yet handle mains specifically, I suggest to merge this PR in the meantime, having:

- no `lights:state/on` property (disable switch)
- `"type": "$TYPE_SMART_PLUG"`, because it currently fits better with the Phoscon GUI.

The remarks by @BabaIsYou for the similar devices (`"_TZ3000_ky0fq4ho","_TZ3000_8bxrzyxz"`) are in this PR too, though untested by myself.

Please DEVs, review/change the timings to your likings. Thanks.

If this gets merged, then https://github.com/dresden-elektronik/deconz-rest-plugin/pull/7012 is not valid anymore.

cheers