Fix Ikea scenes with on/off going wild
Issue: https://github.com/dresden-elektronik/deconz-rest-plugin/issues/8187

When a scene with on=false is recalled, we send a Off with effect command to Ikea lights since they otherwise won't turn off physically. There is also a StateChange added to ensure this works.

Hoewever if a scene is then called with on=true, the StateChange *might* still be alive and turn the light off again. This can be seen in the issue.

The PR clears state/on StateChange for Ikea lights when a scene is recalled preventing old state to be applied on a newer one.