DDF for Namron 8 button switch (4512772, 4512773) (#8240)
This adds support for the Namron 8 button switches 4512772 and 4512773

I have successfully tested that this works for 4512773.

The reason for the new block in button_maps.json is that:

The dim up buttons (1,3,4,5) send these commands:
When holding down: MOVE_WITH_ON_OFF (0x05)
When releasing: STOP_WITH_ON_OFF (0x07)

The dim down buttons (2,4,6,8) send these commands:
When holding down: MOVE (0x01)
When releasing: STOP (0x03)

What I don't know is if 4512772 (or other devices) also need to use the new block in button_maps.json or not.
(I only own the 4512773)

The capitalization of the manufacturer should be all-caps for both 4512772 and 4512773, but not for the other Namron devices.

This is a fork of #7615 . See the discussion in that PR for more details.