Fix creation of invalid Bindings
Seen on macOS somehow clang did not init all fields of the binding to zero when creation an object via `DDF_Binding result = {}`.
The u64 dstExtAddress wasn't initialized only the u16 dstGroup in the union. The ext address was some random stuff from memory.

The fix ensures all fields and dstExtAdress are zero. And the unicast dst address is filled with the actual coordinator mac address.