Disable legacy /gateways endpoint (#8432)
This is an old unofficial and undocumented API. It was used in a customer project to cascade multiple gateways to forward group commands between them.

However this hasn't been used or maintained in years and the networking code shows errors. Therefore disaple it by default for now (compile time option).